refactor(devtools): rename metadata folder and add devtool skill
