+++ .changeset/silver-eagle-drift.md
@@ -1,0 +1,5 @@
+---
+"@devtools/registry": patch
+---
+
+Rename metadata folder from `team-platform` to `platform`, remove unused `icon` field from `ToolMetadata` schema, and consolidate the add-tool guide into `README.md`
