+++ devtools/registry/src/metadata/index.ts
@@ -1,1 +1,1 @@
-export * from "./team-platform";
+export * from "./platform";
